requirements.txt
Django
djangorestframework
django-filter
python-dotenv
